+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   special_consideration_flaw_missing_cve_description (OSIDB-2955)
 - special_handling_flaw_missing_statement Alert to
   special_consideration_flaw_missing_statement (OSIDB-2955)
+- Allow setting empty impact value on flaw (OSIDB-3128)
 
 ### Fixed
 - Fix duplicate comment issue leading in internal server error (OSIDB-3086)